rename : utils 구조화 및 테스트 코드 점검
