+++ client/src/shared/utils/check/hasEmpty.test.ts
