+++ client/src/shared/utils/check/hasEmpty.ts
