+++ client/src/shared/utils/cookies/cookiesManager.test.ts
