+++ client/src/shared/utils/cookies/cookiesManager.ts
