+++ client/src/shared/utils/index.ts
@@ -1,9 +1,9 @@
-export { METHOD, fetcher } from './fetcher';
-export * as hasEmpty from './hasEmpty';
-export * as formatTime from './formatTime';
-export { default as getServerCookie } from './getServerCookies';
-export { default as keys } from './keys';
-export { default as runValidationChecks } from './runValidationChecks';
-export { default as SessionStorageManager } from './sessionStorageManager';
-export { default as LocalStorageManager } from './localStorageManager';
-export * from './cookiesManager';
+export { METHOD, fetcher } from './format/fetcher';
+export * as hasEmpty from './check/hasEmpty';
+export * as formatTime from './format/formatTime';
+export { default as getServerCookie } from './cookies/getServerCookies';
+export { default as keys } from './format/keys';
+export { default as runValidationChecks } from './check/runValidationChecks';
+export { default as SessionStorageManager } from './storage/sessionStorageManager';
+export { default as LocalStorageManager } from './storage/localStorageManager';
+export * from './cookies/cookiesManager';
