+++ client/src/shared/utils/storage/localStorageManager.ts
