+++ client/src/shared/utils/storage/sessionStorageManager.ts
